Add files via upload
A number guessing game aims to guess the number that the program has come up with. Essentially the program logic is:
The program randomly selects a number between 1 and 100 or any other combination of numbers.
It will then ask the player to enter his proposal.
It will then check if this number is the same as the one generated randomly by the computer; if so, the player wins.
If the player’s guess is not the same, then he will check if the number is higher or lower than the guess and tell the player.